Register missing IMessageBoxViewModel in DI, organize MessageBox-related classes, add FakeMessageBoxViewModel, fix resolution of current Application in AvaloniaDispatcher, and remove unused dependency on libdgiplus.
