+++ Desktop.Linux/Services/FileTransferServiceLinux.cs
@@ -3,6 +3,7 @@ using Immense.RemoteControl.Desktop.Shared.Abstractions;
 using Immense.RemoteControl.Desktop.Shared.Services;
 using Immense.RemoteControl.Desktop.Shared.ViewModels;
 using Immense.RemoteControl.Desktop.UI.Controls;
+using Immense.RemoteControl.Desktop.UI.Controls.Dialogs;
 using Immense.RemoteControl.Desktop.UI.Services;
 using Immense.RemoteControl.Desktop.UI.Views;
 using Microsoft.Extensions.Logging;
