+++ Desktop.UI/Controls/Dialogs/MessageBox.axaml
@@ -5,11 +5,15 @@
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
         xmlns:vm="clr-namespace:Immense.RemoteControl.Desktop.UI.ViewModels"
 		xmlns:fakes="clr-namespace:Immense.RemoteControl.Desktop.UI.ViewModels.Fakes"
-        x:Class="Immense.RemoteControl.Desktop.UI.Controls.MessageBox"
+        x:Class="Immense.RemoteControl.Desktop.UI.Controls.Dialogs.MessageBox"
         Icon="{Binding WindowIcon}"
         Title="{Binding Caption}" 
         SizeToContent="WidthAndHeight" MinWidth="300" MinHeight="100" 
         WindowStartupLocation="CenterScreen">
+
+	<Design.DataContext>
+		<fakes:FakeMessageBoxViewModel />
+	</Design.DataContext>
 
   <StackPanel Margin="10, 20">
     <TextBlock Text="{Binding Message}"></TextBlock>
