+++ Desktop.UI/Controls/Dialogs/MessageBoxResult.cs
@@ -1,0 +1,9 @@
+﻿namespace Immense.RemoteControl.Desktop.UI.Controls.Dialogs;
+
+public enum MessageBoxResult
+{
+    Cancel,
+    OK,
+    Yes,
+    No
+}
