+++ Desktop.UI/Controls/Dialogs/MessageBoxType.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Immense.RemoteControl.Desktop.UI.Controls.Dialogs;
+
+public enum MessageBoxType
+{
+    OK,
+    YesNo
+}
